ubacen error-handling za faktore i kategorije, ubacen toastr za informisanje o uspesnosti operacija, omoguceno kreiranje i dodavanje novih stavki cenovnika u aktuelni cenovnik

+++ dataAccess/src/main/java/da/categoryFactor/CategoryFactorServiceImpl.java
@@ -1,11 +1,16 @@
 package da.categoryFactor;
 
 import java.util.List;
+import java.util.Set;
 
 import javax.transaction.Transactional;
+import javax.ws.rs.BadRequestException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import da.factor.Factor;
+import da.factor.FactorRepository;
 
 @Service
 @Transactional
@@ -13,6 +18,9 @@ public class CategoryFactorServiceImpl implements CategoryFactorService{
 	
 	@Autowired
 	private CategoryFactorRepository repository;
+
+	@Autowired
+	private FactorRepository factorRepository;
 
 	@Override
 	public List<CategoryFactor> findAll() {
@@ -31,7 +39,15 @@ public class CategoryFactorServiceImpl implements CategoryFactorService{
 
 	@Override
 	public void delete(Long id) {
-		repository.deleteById(id);
+		/* treba spustiti na repo*/
+		boolean temp = false;
+		List<Factor> list = (List<Factor>)factorRepository.findAll();
+		for (Factor factor : list) 
+			if(factor.getCategory().getId() == id)
+					temp = true;
+		if(temp == false)
+			repository.deleteById(id);
+		else
+			throw new BadRequestException();
 	}
-
 }
