+++ dataAccess/src/main/java/da/factor/FactorServiceImpl.java
@@ -1,11 +1,17 @@
 package da.factor;
 
 import java.util.List;
+import java.util.Set;
 
 import javax.transaction.Transactional;
+import javax.ws.rs.BadRequestException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import da.insurancePolicy.InsurancePolicy;
+import da.insurancePolicy.InsurancePolicyRepository;
+import da.priceListItem.PriceListItem;
 
 @Service
 @Transactional
@@ -13,6 +19,9 @@ public class FactorServiceImpl implements FactorService {
 
 	@Autowired
 	private FactorRepository repository;
+
+	@Autowired
+	private InsurancePolicyRepository incusranceRepository;
 
 	@Override
 	public List<Factor> findAll() {
@@ -31,7 +40,20 @@ public class FactorServiceImpl implements FactorService {
 
 	@Override
 	public void delete(Long id) {
-		repository.deleteById(id);
+		/* treba spustiti na repo*/
+		boolean temp = false;
+		List<InsurancePolicy> list = (List<InsurancePolicy>)incusranceRepository.findAll();
+		for (InsurancePolicy insurancePolicy : list) {
+			Set<PriceListItem> priceListItems= insurancePolicy.getPriceListItems();
+			for (PriceListItem priceListItem : priceListItems) {
+				if(priceListItem.getFactor().getId()==id)
+					temp = true;
+			}
+		}
+		if(temp == false)
+			repository.deleteById(id);
+		else 
+			throw new BadRequestException();
 	}
 
 	@Override
